#124 fix test
