+++ unit_test/test_mock_telescope.py
@@ -1,7 +1,8 @@
 import unittest
 import config
 import socket
-from unittest.mock import MagicMock
+import datetime
+from unittest.mock import DEFAULT, MagicMock
 from components.telescope.simulator.telescope import Telescope
 import socket
 from base.singleton import Singleton
@@ -14,6 +15,9 @@ class MockTelescopeTest(unittest.TestCase):
         self.telescopio = Telescope()
         socket.socket.connect = MagicMock(return_value=None)
         socket.socket.close = MagicMock(return_value=None)
+        self.configGetInt = config.Config.getInt
+        self.configGetFloat = config.Config.getFloat
+        self.configGetValue = config.Config.getValue
 
     def tearDown(self):
         alt = self.telescopio.park_alt
@@ -39,3 +43,38 @@ class MockTelescopeTest(unittest.TestCase):
     def test_move_tele(self):
         self.telescopio.move_tele(tr=0, az=0, alt=0)
         self.assertEqual(self.telescopio.coords, {"tr": 0, "alt": 0, "az": 0, "error": 0})
+
+    def __side_effect_config__(self, key, section=""):
+        if key == "park_alt" and section == "telescope":
+            side_effect = 0.2
+        elif key == "park_az" and section == "telescope":
+            side_effect = 0.1
+        elif key == "lat" and section == "geography":
+            side_effect = "+12d48.69m"
+        elif key == "lon" and section == "geography":
+            side_effect = "42d13.76m"
+        elif key == "height" and section == "geography":
+            side_effect = 465
+        elif key == "name_obs" and section == "geography":
+            side_effect = "157FrassoSabino"
+        elif key == "equinox" and section == "geography":
+            side_effect = "J2000"
+        elif key == "timezone" and section == "geography":
+            side_effect = "Europe/Rome"
+        else:
+            side_effect = DEFAULT
+        return side_effect
+
+    def test_conv_altaz_to_ardec(self):
+        date = datetime.datetime(2020, 12, 6, 15, 29, 43, 79060, tzinfo=datetime.timezone.utc)
+        config.Config.getInt = MagicMock(side_effect=self.__side_effect_config__)
+        config.Config.getFloat = MagicMock(side_effect=self.__side_effect_config__)
+        config.Config.getValue = MagicMock(side_effect=self.__side_effect_config__)
+        coords = self.telescopio.sync(date)
+        self.assertEqual(coords["ar"], 9.364517932878291)
+        self.assertEqual(coords["dec"], 47.96211142851545)
+
+    def tearDown(self):
+        config.Config.getInt = self.configGetInt
+        config.Config.getFloat = self.configGetFloat
+        config.Config.getValue = self.configGetValue
